Conflicts:
	Engine/include/Window/GameWindow.hpp
	Engine/src/Core/Engine.cpp
	Engine/src/Window/GameWindow.cpp
	Game/include/GameStates/PlayState.hpp
	Game/src/GameStates/PlayState.cpp

+++ Engine/src/Utils/Timer.cpp
@@ -1,5 +1,6 @@
-#include "Utils/Timer.hpp"
 #include <GLFW/glfw3.h>
+
+#include "Utils/Timer.hpp"
 
 Timer::Timer()
 {
@@ -10,10 +11,10 @@ Timer::~Timer() {}
 
 void    Timer::reset()
 {
-    _lastReset = glfwGetTime();
+    _lastReset = (float)glfwGetTime();
 }
 
-double  Timer::getElapsedTime() const
+float   Timer::getElapsedTime() const
 {
-    return (glfwGetTime() - _lastReset);
+    return ((float)glfwGetTime() - _lastReset);
 }
